+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   * Thanks to @Manishearth for identifying this.
 * ci: Check the tests pass with `-Zminimal-versions`
 
+### Added
+* Implement slog::Value for `std::sync::LazyLock` and `std::cell::LazyCell`.
+
 ## 2.8.1 - 2025-10-05
 This fixes an accidental breaking change in the v2.8.0 release,
 where the public API was changed from `erased_serde v0.4` to `erased_serde v0.3`.